add: validation mechanism for ContactRequest payload in the api hanlder.
