+++ packages/types/src/validation.ts
@@ -62,4 +62,9 @@ export enum FieldError {
   InvalidBankName = "invalid-bank-name",
   ShortTopicName = "short-topic-name",
   LongTopicName = "long-topic-name",
+  ShortContactRequestTitle = "short-contact-request-title",
+  LongContactRequestTitle = "long-contact-request-title",
+  ShortContactRequestMessage = "short-contact-request-message",
+  LongContactRequestMessage = "long-contact-request-message",
+  InvalidType = "invalid-type",
 }
