+++ packages/ui/src/constants/error.ts
@@ -118,4 +118,13 @@ export const apiErrorMap: Record<ApiErrorCode, LocalId> = {
   [FieldError.InvalidBankName]: "error.field.invalid-bank-name",
   [FieldError.ShortTopicName]: "error.field.short-topic-name",
   [FieldError.LongTopicName]: "error.field.long-topic-name",
+  [FieldError.InvalidType]: "error.field.invalid-type",
+  [FieldError.ShortContactRequestTitle]:
+    "error.field.short-contact-request-title",
+  [FieldError.LongContactRequestTitle]:
+    "error.field.long-contact-request-title",
+  [FieldError.ShortContactRequestMessage]:
+    "error.field.short-contact-request-message",
+  [FieldError.LongContactRequestMessage]:
+    "error.field.long-contact-request-message",
 };
