+++ packages/utils/package.json
@@ -89,9 +89,9 @@
       "import": "./dist/esm/withdrawMethod.js",
       "require": "./dist/cjs/withdrawMethod.js"
     },
-    "./verification": {
-      "import": "./dist/esm/verification.js",
-      "require": "./dist/cjs/verification.js"
+    "./validation": {
+      "import": "./dist/esm/validation/index.js",
+      "require": "./dist/cjs/validation/index.js"
     },
     "./session": {
       "import": "./dist/esm/session.js",
@@ -117,7 +117,7 @@
     "./utils": "./dist/cjs/utils.js",
     "./value": "./dist/cjs/value.js",
     "./withdrawMethod": "./dist/cjs/withdrawMethod.js",
-    "./verification": "./dist/cjs/verification.js"
+    "./validation": "./dist/cjs/validation/index.js"
   },
   "devDependencies": {
     "@faker-js/faker": "^9.0.3",
